% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npxplore.R
\name{npx_anova}
\alias{npx_anova}
\title{Calculate multi-condition differential proteins with one-way ANOVA}
\usage{
npx_anova(
  data,
  variable = "Condition",
  covariates = NULL,
  check_log = NULL,
  logFC = 0,
  adj.p = 0.05
)
}
\arguments{
\item{data}{Prepared assay data.}

\item{variable}{Metadata column used as the grouping variable.}

\item{covariates}{Optional covariate column names for ANOVA.}

\item{check_log}{Optional Olink QC log.}

\item{logFC}{Minimum absolute Log2 fold-change for DEPs.}

\item{adj.p}{Maximum adjusted p-value for DEPs.}
}
\value{
A list containing the complete result in \code{result} and filtered DEPs in \code{deps}.
}
\description{
Calculate multi-condition differential proteins with one-way ANOVA
}
