% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npxplore.R
\name{npx_anova_posthoc}
\alias{npx_anova_posthoc}
\title{Calculate ANOVA post-hoc contrasts}
\usage{
npx_anova_posthoc(
  data,
  analysis = NULL,
  variable = "Condition",
  covariates = NULL,
  check_log = NULL,
  logFC = 0,
  adj.p = 0.05
)
}
\arguments{
\item{data}{Prepared assay data.}

\item{analysis}{An object returned by \code{\link[=npx_analyze]{npx_analyze()}}.}

\item{variable}{Metadata grouping column.}

\item{covariates}{Optional ANOVA covariates.}

\item{check_log}{Optional Olink QC log.}

\item{logFC}{Minimum absolute post-hoc Log2 fold-change.}

\item{adj.p}{Maximum adjusted post-hoc p-value.}
}
\value{
A filtered post-hoc result table.
}
\description{
Calculate ANOVA post-hoc contrasts
}
