% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npxplore.R
\name{npx_prepare_analysis}
\alias{npx_prepare_analysis}
\title{Prepare grouping levels and factor order for differential analysis}
\usage{
npx_prepare_analysis(data, variable = "Condition", min_levels = 2L)
}
\description{
Prepare grouping levels and factor order for differential analysis
}
\keyword{internal}
