% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npxplore.R
\name{npx_prepare_two_group_data}
\alias{npx_prepare_two_group_data}
\title{Keep assays with usable observations in both two-condition groups}
\usage{
npx_prepare_two_group_data(data, variable)
}
\description{
Keep assays with usable observations in both two-condition groups
}
\keyword{internal}
