% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npxplore.R
\name{npx_ttest}
\alias{npx_ttest}
\title{Calculate two-condition differential proteins with an Olink t-test}
\usage{
npx_ttest(
  data,
  variable = "Condition",
  pair_id = NULL,
  check_log = NULL,
  logFC = 0,
  adj.p = 0.05
)
}
\arguments{
\item{data}{Prepared assay data.}

\item{variable}{Metadata column used as the grouping variable.}

\item{pair_id}{Optional pairing identifier for paired tests.}

\item{check_log}{Optional Olink QC log.}

\item{logFC}{Minimum absolute Log2 fold-change for DEPs.}

\item{adj.p}{Maximum adjusted p-value for DEPs.}
}
\value{
A list containing the complete result in \code{result} and filtered DEPs in \code{deps}.
}
\description{
Calculate two-condition differential proteins with an Olink t-test
}
