% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npxplore.R
\name{npx_validate_inputs}
\alias{npx_validate_inputs}
\title{Validate NPX data and metadata}
\usage{
npx_validate_inputs(data, meta, sample_id_col = "SampleID")
}
\arguments{
\item{data}{An Olink NPX data frame.}

\item{meta}{A metadata data frame.}

\item{sample_id_col}{Metadata sample identifier column.}
}
\description{
Validate NPX data and metadata
}
