% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npxplore.R
\name{npx_wilcon}
\alias{npx_wilcon}
\title{Backward-compatible alias for \code{\link[=npx_wilcox]{npx_wilcox()}}.}
\usage{
npx_wilcon(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=npx_wilcox]{npx_wilcox()}}.}
}
\description{
Backward-compatible alias for \code{\link[=npx_wilcox]{npx_wilcox()}}.
}
