% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npxplore.R
\name{print.npxplore_result}
\alias{print.npxplore_result}
\title{Print an NPXplore result}
\usage{
\method{print}{npxplore_result}(x, ...)
}
\arguments{
\item{x}{An \code{npxplore_result} object.}

\item{...}{Unused additional arguments.}
}
\description{
Print an NPXplore result
}
